# Conflicts:
#	main-service/src/main/java/ru/practicum/ewm/dao/EventRepository.java
#	main-service/src/main/java/ru/practicum/ewm/service/EventService.java
#	main-service/src/main/java/ru/practicum/ewm/service/EventServiceImpl.java

+++ main-service/src/main/java/ru/practicum/ewm/mapper/CompilationMapper.java
@@ -15,7 +15,7 @@ public class CompilationMapper {
 
     public static CompilationDto toCompilationDto(Compilation compilation,
                                                   Map<Long, Long> viewsMap,
-                                                  Map<Long, Long> confirmedMap) {
+                                                  Map<Long, Integer> confirmedMap) {
         return CompilationDto.builder()
                 .id(compilation.getId())
                 .title(compilation.getTitle())
@@ -28,7 +28,7 @@ public class CompilationMapper {
 
     private static EventShortDto toEventShortDto(Event event,
                                                  Map<Long, Long> viewsMap,
-                                                 Map<Long, Long> confirmedMap) {
+                                                 Map<Long, Integer> confirmedMap) {
         Long eventId = event.getId();
 
         return EventShortDto.builder()
@@ -40,7 +40,7 @@ public class CompilationMapper {
                 .category(CategoryMapper.toDto(event.getCategory()))
                 .initiator(new UserShortDto(event.getInitiator().getId(), event.getInitiator().getName()))
                 .views(viewsMap.getOrDefault(eventId, 0L))
-                .confirmedRequests(confirmedMap.getOrDefault(eventId, 0L))
+                .confirmedRequests(confirmedMap.getOrDefault(eventId, 0))
                 .build();
     }
 }
