+++ main-service/src/main/java/ru/practicum/ewm/model/EventShortView.java
@@ -21,5 +21,5 @@ public interface EventShortView {
 
     Boolean getPaid();
 
-    Long getConfirmedRequests();
+    Integer getConfirmedRequests();
 }
